python_src/requirements.txt
playwright==1.40.0
openai==1.6.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
jinja2==3.1.2